Fix emacs native compilation warning
```
 ■  Warning (comp): diminish.el:155:2: Warning: docstring has wrong
usage of unescaped single quotes (use \= or different quoting)
 ■  Warning (comp): diminish.el:207:2: Warning: docstring has wrong
usage of unescaped single quotes (use \= or different quoting)
```

To reproduce, have emacs build with native compilation and notice the
compilation logs. You can then open this file and run `M-x
emacs-lisp-native-compile-and-load` before and after the changes to see the
warning is removed.